requirements.txt
absl-py==2.1.0
addict==2.4.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
einops==0.7.0
filelock==3.13.1
fonttools==4.50.0
future==1.0.0
gdown==5.1.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
imageio==2.34.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
joblib==1.3.2
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.2
lazy_loader==0.3
lmdb==1.4.1
lpips==0.1.4
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.7.5
natsort==8.4.0
networkx==3.1
numpy==1.24.1
oauthlib==3.2.2
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.2.0
platformdirs==4.2.0
protobuf==5.26.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
pyqt5-tools==5.15.9.3.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyWavelets==1.4.1
PyYAML==6.0.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soupsieve==2.5
tb-nightly==2.14.0a20230808
tensorboard-data-server==0.7.2
threadpoolctl==3.4.0
tifffile==2023.7.10
tomli==2.0.1
torch==1.10.2+cu111
torchaudio==0.10.2+cu111
torchvision==0.11.3+cu111
tqdm==4.66.2
typing_extensions==4.8.0
urllib3==2.2.1
Werkzeug==3.0.1
yacs==0.1.8
yapf==0.40.2
zipp==3.18.1
